pyproject.toml
[project]
name = "newsletter-autoresearch"
version = "0.1.0"
description = "Self-improving newsletter optimization system inspired by Karpathy's autoresearch"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
]
